requirements.txt
# Stable, battle-tested dependencies - no bullshit versions
fastapi==0.68.0
uvicorn[standard]==0.15.0
sqlalchemy[asyncio]==1.4.23
alembic==1.7.1
pydantic==1.8.2
python-multipart==0.0.5
python-dotenv==0.19.0
psycopg2-binary==2.9.1
asyncpg==0.24.0
aiofiles==0.7.0

# Optional for full functionality
# python-telegram-bot==13.7
# redis==3.5.3
# python-jose==3.3.0
# passlib==1.7.4
